+++ source/Relativity.Testing.Framework.Api/Strategies/Users/V1/UserCreateStrategyV1.cs
@@ -11,8 +11,9 @@ namespace Relativity.Testing.Framework.Api.Strategies
 			IRestService restService,
 			IRequireStrategy<Client> clientRequireStrategy,
 			IChoiceResolveByObjectFieldAndNameStrategy choiceResolveByObjectFieldAndNameStrategy,
-			IUserAddToGroupStrategy userAddToGroupStrategy)
-			: base(restService, clientRequireStrategy, choiceResolveByObjectFieldAndNameStrategy, userAddToGroupStrategy)
+			IUserAddToGroupStrategy userAddToGroupStrategy,
+			IUserSetPasswordStrategy userSetPasswordStrategy)
+			: base(restService, clientRequireStrategy, choiceResolveByObjectFieldAndNameStrategy, userAddToGroupStrategy, userSetPasswordStrategy)
 		{
 		}