+++ CHANGELOG.md
@@ -5,11 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [6.1.4] - 2022-04-06
+## [6.1.6] - 2022-04-15
 
 ### Updated
 
 - Updated the supported version range to 13.2.
+
+## [6.1.5] - 2022-04-12
+
+### Changed
+
+- WaitDeleteWorkspaceStrategy now has 3 minutes timeout.
+
+## [6.1.4] - 2022-04-11
+
+### Changed
+
+- IAccountPoolService now resets the password of existing accounts to match the expected password.
 
 ## [6.1.3] - 2022-02-17